+++ neo/d3xp/gamesys/SaveGame.cpp
@@ -300,7 +300,7 @@ void idSaveGame::WriteString( const char *string ) {
 	stringHash.Add( hash, stringTable.Num() - 1 );
 
 	WriteInt( curStringTableOffset );
-	curStringTableOffset += ( strlen( string ) + 4 );
+	curStringTableOffset += ( static_cast<int>(strlen( string )) + 4 );
 }
 
 /*
